requirements.txt
py-cord==2.0.0
snowflake-id==0.0.2
python-dotenv==0.20.0
python-Levenshtein==0.12.2
fuzzywuzzy==0.18.0
requests==2.28.1
dataclasses==0.6
dataclasses-json==0.5.7
